update language model
